refactor: split ledger tests and message tool options
